Move the Docker commands and use a machine
