--- src/Cli/Application.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Dock\Cli;
-
-use Symfony\Component\Console\Application as ConsoleApplication;
-
-class Application extends ConsoleApplication
-{
-}
