+++ src/Docker/Machine/Machine.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Dock\Docker\Machine;
+
+interface Machine
+{
+    public function isRunning();
+    public function start();
+    public function stop();
+}
